Add files via upload
This python script (1) iterates through your Mosyle assets, (2) pulls the serial of each device in Mosyle, (3) checks to see if that serial is in Snipe-IT, and (4) if it is in Snipe-IT, custom-field(s) in Snipe-IT is(are) updated. Specifically, I wrote this script so we could easily find what MacOS assets are not bound to MDM yet. From Snipe-IT now, it is easy to pull a list of all devices not bound so we can more easily track them down and bind them. 
-----
Lines 5-7 require input. Add your Snipe-IT URL, Mosyle API token, and Snipe-IT API token. 
---
Beginning at line 23, you will need to replace the payload information with your Snipe-IT DB information (see here how to find: https://snipe-it.readme.io/reference#updating-custom-fields).... Replace the DB info with your DB info, and replace the "Bound" and "Mosyle" values to correspond to whatever value you want to modify specifically in Snipe-IT.

Contact me with questions.